ステップ1と2のRspec作成
