+++ app/views/tasks/index.html.erb
@@ -8,7 +8,7 @@
   </tr>
 
 <% @tasks.each do |task| %>
-  <tr>
+  <tr class="task_row">
     <td><%= task.title %></td>
     <td><%= task.content %></td>
