+++ package.json
@@ -12,6 +12,7 @@
   },
   "version": "0.1.0",
   "devDependencies": {
+    "puppeteer": "^16.0.0",
     "webpack-dev-server": "^3"
   }
 }
